ResourceSetを分離
